+++ src/main/java/org/psidnell/omnifocus/Main.java
@@ -110,7 +110,7 @@ public class Main extends CommandLine {
 
     public static void main(String[] args) throws Exception {
 
-        ApplicationContext appContext = ApplicationContextFactory.create();
+        ApplicationContext appContext = ApplicationContextFactory.getContext();
 
         Main main = appContext.getBean("main", Main.class);